experiments/windows/scripts/requirements.txt
google-cloud-secret-manager>=2.16
